==> notebooks/1.0-wfp-etl-process.ipynb <==
# ETL Process to clean data before analysis

## Preparing environment

import pandas as pd
import numpy as np
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer
from sklearn.ensemble import RandomForestRegressor
from sklearn.preprocessing import LabelEncoder

import sys
sys.path.append('../high_performance_employee_resign_prediction')
from utils import paths

## Data Dictionary

* id_employee - Employee ID
* id_last_boss - Boss ID
* seniority - 1: For people who don't have anyone in charge, 2: For those who have people in charge
* work_modality - Work modality specified on employee contract
* office_distance -  Distance in kilometers from employee home to work.
* low_health_days -  Number of days pf justified sick leave since employee started to work for the company.
* gender - Employee gender
* recruitment_channel - Indicates the means by which the application process took place. 
* average_permanence - Mean time in years that the employee has worked in previous laboral experience.
* birth_date - Birth date in format dd/mm/yyyy.
* salary - It is the amount corresponding to the monthly remuneration received.
* psi_score - Is the score obtained in psychometric test during selection process.
* join_date - Date when the employee started to work for the company.
* marital_estatus - Marital Estatus.
* performance_score - Is the score obtained by the employee it his last quarterly performance test if he had.
* resign - "0" if employee continues working or "1" if employee left the company on his first 6 months.

## Extracting data

train_df = pd.read_csv(paths.data_raw_dir('train_data.csv'), parse_dates=['birth_date', 'join_date'], dayfirst=True, sep=';')

test_df = pd.read_csv(paths.data_raw_dir('test_data.csv'), parse_dates=['birth_date', 'join_date'], dayfirst=True, sep=';')

## Checking data

train_df.info()

test_df.info()

train_df.describe()

test_df.describe()

Observing the statistics summary, we can see that there are some outliers in `office_distance`, `low_health_days`, `average_permanence` and `salary` for both train and test data. They will be analyzed and actions will be taken during the EDA.

## Looking for typo errors in object data

train_df.describe(include='object')

test_df.describe(include='object')

# Checking work_modality categories

print(train_df.work_modality.unique())
print(test_df.work_modality.unique())

# Checking recruitment_channel categories

print(train_df.recruitment_channel.unique())
print(test_df.recruitment_channel.unique())

# Checking marital_estatus categories

print(train_df.marital_estatus.unique())
print(test_df.marital_estatus.unique())

From the data dictionary we can see that `seniority` and `id_last_boss` are categories too. However, id_last_boss has many categories, so it will be analyzed during the hypothesis testing phase.

cat_cols = ['work_modality', 'gender', 'recruitment_channel', 'marital_estatus', 'seniority']

# Defining a function to check categories proportion

def check_categories_proportion(df, col):
    
    """
    Calculate and print the proportion of each category in a specified column of a DataFrame.

    This function takes a pandas DataFrame and a column name, calculates the proportion 
    (as a percentage) of each unique value (category) in the specified column, and prints 
    the resulting proportions.

    Parameters:
    df (pandas.DataFrame): The DataFrame containing the data.
    col (str): The name of the column for which to calculate category proportions.

    Returns:
    None: This function prints the category proportions and does not return any value.

    Example:
    >>> import pandas as pd
    >>> data = {'category': ['A', 'B', 'A', 'C', 'B', 'A']}
    >>> df = pd.DataFrame(data)
    >>> check_categories_proportion(df, 'category')
    A    50.0
    B    33.3
    C    16.7
    Name: category, dtype: float64
    """
    
    proportion = df[col].value_counts(normalize=True) * 100
    print(proportion)

for col in cat_cols:
    check_categories_proportion(train_df, col)
    print('-'*40)

for col in cat_cols:
    check_categories_proportion(test_df, col)
    print('-'*40)

Categories proportion in both train and test data are similar.

## Extracting info from date columns

# Calculating age of join

train_df['join_age'] = (train_df['join_date'] - train_df['birth_date']).dt.days // 365

test_df['join_age'] = (test_df['join_date'] - test_df['birth_date']).dt.days // 365

# Obtaining season info

train_df['join_year'] = train_df['join_date'].dt.year
train_df['join_month'] = train_df['join_date'].dt.month

test_df['join_year'] = test_df['join_date'].dt.year
test_df['join_month'] = test_df['join_date'].dt.month

train_df.info()

test_df.info()

## Missing values treatment

Before imputing missing values, lets explore them by concatenating both DataFrames. From dataframe info is known that there are null values on id_last_boss in both train and test datasets, and on performance score in train dataset.

Additionally, since bosses are employees of the company as well, they are included in the employee list and they have seniority 2.

# Concatenating both dataframes since they have similar characteristics

concat_df = pd.concat([train_df.drop('resign', axis=1), test_df], axis=0).reset_index(drop=True)

# Checking null values per row

concat_df.isna().sum(axis=1).value_counts().to_frame()

There are only two rows where there are 2 missing values, and 238 where there are 238. Let's explore null values on id_last_boss

# Exploring id_last_boss null values

concat_df[concat_df['id_last_boss'].isna()]

# Checking the values of seniority from these employees

concat_df[concat_df['id_last_boss'].isna()].seniority.value_counts()

All the seniority values are 2, which means that all these employees are bosses that lead their own division and don't have any superior. This means that missing values can be imputed with their own IDs.

concat_df['id_last_boss'].fillna(concat_df['id_employee'], inplace=True)

concat_df[concat_df['seniority'] == 2]

# Checking null values per row again

concat_df.isna().sum(axis=1).value_counts().to_frame()

Since there are few missing values, we can impute them using an imputation method. In order to preserve the distribution of the sample, we can use an IterativeImputer with a RandomForestRegressor. 

# Defining categorical and numerical columns

num_cols = ['office_distance', 'low_health_days', 'average_permanence', 'salary', 'performance_score', 'psi_score', 'join_age']
cat_cols = ['id_last_boss', 'seniority', 'work_modality', 'gender', 'recruitment_channel', 'marital_estatus']

# join_year and join_month are categories too but they are already encoded

# Saving features that will not be used during the imputation

feat_not_used = ['id_employee', 'birth_date', 'join_date']

concat_fnu = concat_df[feat_not_used]

y = train_df[['resign']]

# Converting id_last_boss on it dtype

concat_df['id_last_boss'] = concat_df['id_last_boss'].astype(int)

# Getting only features to execute imputation

X = concat_df.drop(feat_not_used, axis=1)

# Initializing LabelEncoder

label_encoders = {}

# Applying LabelEncoder to each categorical column

for col in cat_cols:
    le = LabelEncoder()
    X[col] = le.fit_transform(X[col].astype(str))
    label_encoders[col] = le

X.head()

# Initializing the IterativeImputer with RandomForestRegressor as the estimator
imputer = IterativeImputer(estimator=RandomForestRegressor(), max_iter=20, min_value=X.performance_score.min(),
                           max_value=X.performance_score.max(), random_state=42)

# Impute missing values
X_imputed = pd.DataFrame(imputer.fit_transform(X), columns=X.columns)

X_imputed.info()

No missing values anymore

X_imputed.head()

# Converting back the features to integers

for col in X_imputed.columns:
    if col != 'office_distance':
        X_imputed[col] = X_imputed[col].astype(int)

X_imputed.head()

# Decoding all the columns

for col in cat_cols:
    X_imputed[col] = label_encoders[col].inverse_transform(X_imputed[col])
    
X_imputed.head()

# Getting the non used features

concat_df_final = pd.concat([concat_fnu, X_imputed], axis=1)
concat_df_final.head()

concat_df_final.info()

## Creating new features

# Creating a performance column to determine if a employee has high or low perfomance
concat_df_final['performance'] = ['high' if score >= 80 else 'low' for score in concat_df_final['performance_score']]

# Creating an age-group column
concat_df_final['join_age_group'] = ['young' if age <= 26 else ('adult' if age <= 59 else 'old_adult') for age in concat_df_final['join_age']]

concat_df_final.head()

## Joining dataframe to create new features

Boss ids are also present in employee id, so we can use this information to create new features that can help to improve the results

# Changing id_last_boss dtype
concat_df_final['id_last_boss'] = concat_df_final['id_last_boss'].astype(int)

expanded_df = pd.merge(concat_df_final, concat_df_final, how='left',
                             left_on='id_last_boss', right_on='id_employee',
                             suffixes=('', '_boss'))

expanded_df.info()

expanded_df.head()

# Dropping unnecessary features

expanded_df.drop(columns=['id_employee_boss', # Is the same id_last_boss
                          'id_last_boss_boss', # Is the same id_last_boss
                          'seniority_boss', # It's 2 for all bosses
                          ], inplace=True)

## Creating new features

These new features allow us to make new features such as differences in numerical columns, determine if a employee joined after or before a boss and joined, etc

# Creating difference columns

expanded_df['salary_diff'] = expanded_df['salary_boss'] - expanded_df['salary']
expanded_df['join_days_diff'] = (expanded_df['join_date_boss'] - expanded_df['join_date']).dt.days
expanded_df['joined_after_boss'] = [0 if value > 0 else 1 for value in expanded_df['join_days_diff']]
expanded_df['age_diff'] = (expanded_df['birth_date_boss'] - expanded_df['birth_date']).dt.days // 365
expanded_df['younger_than_boss'] = [0 if value > 0 else 1 for value in expanded_df['age_diff']]

# Dropping date columns

expanded_df.drop(columns=['join_date', 'join_date_boss', 'birth_date_boss', 'birth_date'], inplace=True)

# Catching average numerical features of employee per boss

expanded_df['avg_od_epb'] = expanded_df.groupby('id_last_boss')['office_distance'].transform('mean')
expanded_df['avg_lhd_epb'] = expanded_df.groupby('id_last_boss')['low_health_days'].transform('mean')
expanded_df['avg_avgp_epb'] = expanded_df.groupby('id_last_boss')['average_permanence'].transform('mean')
expanded_df['avg_sal_epb'] = expanded_df.groupby('id_last_boss')['salary'].transform('mean')
expanded_df['avg_ps_epb'] = expanded_df.groupby('id_last_boss')['performance_score'].transform('mean')
expanded_df['avg_psis_epb'] = expanded_df.groupby('id_last_boss')['psi_score'].transform('mean')
expanded_df['avg_ja_epb'] = expanded_df.groupby('id_last_boss')['join_age'].transform('mean')

# Getting total employees by boss (minus 1 because bosses have their same id on their id_last_boss)

freq_enc_boss = expanded_df['id_last_boss'].value_counts() - 1

expanded_df['boss_employees_in_charge'] = expanded_df['id_last_boss'].map(freq_enc_boss)

# Separating train and test df

train_final = pd.concat([expanded_df.loc[:2151, :], y], axis=1)
test_final = expanded_df.loc[2152:, :].reset_index(drop=True)

train_final.info()

test_final.info()

## Saving clean data

train_final.to_csv(paths.data_interim_dir('train_clean.csv'), index=False, sep=',')

test_final.to_csv(paths.data_interim_dir('test_clean.csv'), index=False, sep=',')
