==> notebooks/5.0-wfp-feature-engineering.ipynb <==
# Feature engineering

There are libraries to make auto feature engineering. In this case, we gonna use AutoFeat that is a powerful tool to make feature engineering. It requires preprocessing before using it. 

## Preparing environment

import numpy as np
import pandas as pd
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from autofeat import AutoFeatClassifier
import sys
sys.path.append('../high_performance_employee_resign_prediction')
from utils import paths

## Importing data

train_df = pd.read_csv(paths.data_interim_dir('train_node.csv'))
test_df = pd.read_csv(paths.data_interim_dir('test_node.csv'))

num_cols = ['office_distance', 'low_health_days', 'average_permanence', 'salary', 'performance_score', 
            'psi_score', 'join_age', 'office_distance_boss', 'low_health_days_boss', 'average_permanence_boss', 
            'salary_boss', 'performance_score_boss', 'psi_score_boss', 'join_age_boss', 'salary_diff',
            'join_days_diff', 'age_diff', 'avg_od_epb', 'avg_lhd_epb', 'avg_avgp_epb', 
            'avg_sal_epb', 'avg_ps_epb', 'avg_psis_epb', 'avg_ja_epb', 'boss_employees_in_charge']

cat_cols = ['seniority', 'work_modality', 'gender', 'recruitment_channel', 'marital_estatus', 
            'join_year', 'join_month', 'performance', 'join_age_group', 'work_modality_boss', 
            'gender_boss', 'recruitment_channel_boss', 'marital_estatus_boss', 'join_year_boss', 'join_month_boss', 
            'performance_boss', 'join_age_group_boss', 'joined_after_boss', 'younger_than_boss']

## Dropping categories according to hypothesis testing

reduced_train_df = train_df.drop(columns=['join_month', 'work_modality_boss', 'gender_boss', 'recruitment_channel_boss', 
                                          'marital_estatus_boss', 'join_year_boss', 'join_month_boss', 'join_age_group_boss', 
                                          'younger_than_boss', 'office_distance', 'average_permanence', 'salary', 
                                          'psi_score', 'join_age', 'office_distance_boss', 'average_permanence_boss', 
                                          'salary_boss', 'psi_score_boss', 'join_age_boss', 'age_diff', 
                                          'avg_od_epb', 'avg_lhd_epb', 'avg_sal_epb', 'avg_psis_epb', 
                                          'boss_employees_in_charge'])

reduced_test_df = test_df.drop(columns=['join_month', 'work_modality_boss', 'gender_boss', 'recruitment_channel_boss', 
                                          'marital_estatus_boss', 'join_year_boss', 'join_month_boss', 'join_age_group_boss', 
                                          'younger_than_boss', 'office_distance', 'average_permanence', 'salary', 
                                          'psi_score', 'join_age', 'office_distance_boss', 'average_permanence_boss', 
                                          'salary_boss', 'psi_score_boss', 'join_age_boss', 'age_diff', 
                                          'avg_od_epb', 'avg_lhd_epb', 'avg_sal_epb', 'avg_psis_epb', 
                                          'boss_employees_in_charge'])

## Preprocessing full data

ohe_full = OneHotEncoder(sparse_output=False, drop='first')
ss_full = StandardScaler()

# Encoding categories

# Train
encoded_full_train = pd.DataFrame(ohe_full.fit_transform(train_df[cat_cols]))
encoded_full_train.columns = ohe_full.get_feature_names_out(cat_cols)

# Test
encoded_full_test = pd.DataFrame(ohe_full.transform(test_df[cat_cols]))
encoded_full_test.columns = ohe_full.get_feature_names_out(cat_cols)

# Scaling numericals

# Train
scaled_full_train = pd.DataFrame(ss_full.fit_transform(train_df[num_cols]), columns=num_cols)

# Test
scaled_full_test = pd.DataFrame(ss_full.transform(test_df[num_cols]), columns=num_cols)

# Concatenating dfs

concat_full_train = pd.concat([encoded_full_train, scaled_full_train], axis=1)

concat_full_test = pd.concat([encoded_full_test, scaled_full_test], axis=1)

### Feature engineering

Autofeat is a library that applies multiple numerical transformations to the features in order to find meaningful relationships between features and target, improving the prediction

y = train_df['resign']

feateng_cols = ['seniority_2',
 'work_modality_Presencial',
 'gender_Mujer',
 'recruitment_channel_Headhunter',
 'recruitment_channel_Linkedin',
 'recruitment_channel_Portal Web',
 'recruitment_channel_Referidos',
 'marital_estatus_Divorciado',
 'marital_estatus_Soltero',
 'marital_estatus_Viudo',
 'performance_low',
 'join_age_group_old_adult',
 'join_age_group_young',
 'work_modality_boss_Presencial',
 'gender_boss_Mujer',
 'recruitment_channel_boss_Headhunter',
 'recruitment_channel_boss_Linkedin',
 'recruitment_channel_boss_Portal Web',
 'recruitment_channel_boss_Referidos',
 'marital_estatus_boss_Divorciado',
 'marital_estatus_boss_Soltero',
 'marital_estatus_boss_Viudo',
 'performance_boss_low',
 'join_age_group_boss_old_adult',
 'join_age_group_boss_young',
 'joined_after_boss_1',
 'younger_than_boss_1',
 'office_distance',
 'low_health_days',
 'average_permanence',
 'salary',
 'performance_score',
 'psi_score',
 'join_age',
 'office_distance_boss',
 'low_health_days_boss',
 'average_permanence_boss',
 'salary_boss',
 'performance_score_boss',
 'psi_score_boss',
 'join_age_boss',
 'salary_diff',
 'join_days_diff',
 'age_diff',
 'boss_employees_in_charge']

autofeat = AutoFeatClassifier(n_jobs=-1, verbose=1, feateng_cols=feateng_cols)

X_full_transformed = autofeat.fit_transform(concat_full_train, y)

X_full_transformed_test = autofeat.transform(concat_full_test)

### Obtaining generated columns

# Obtaining generated columns
generated_cols = [col for col in X_full_transformed.columns.tolist() if col not in concat_full_train.columns.tolist()]
generated_train = X_full_transformed[generated_cols]
generated_test = X_full_transformed_test[generated_cols]

## Saving full data

# Dropping categories and numerical columns

# Train
train_full_df = pd.concat([train_df, generated_train], axis=1)

# Test
test_full_df = pd.concat([test_df, generated_test], axis=1)

# Saving these dataframes

train_full_df.to_csv(paths.data_processed_dir('train_featured.csv'), index=False, sep=',')
test_full_df.to_csv(paths.data_processed_dir('test_featured.csv'), index=False, sep=',')

## Saving reduced data

# Train
train_red_df = pd.concat([reduced_train_df, generated_train], axis=1)

# Test
test_red_df = pd.concat([reduced_test_df, generated_test], axis=1)

# Saving these dataframes

train_red_df.to_csv(paths.data_processed_dir('train_red_featured.csv'), index=False, sep=',')
test_red_df.to_csv(paths.data_processed_dir('test_red_featured.csv'), index=False, sep=',')
