==> notebooks/6.0-wfp-modeling-data-xgboost.ipynb <==
# Modeling full data

Since our full data has many features due to node embeddings, we need to use robust models as XGBoost and a Neural Network. The metric chosen metric for this evaluation is F1-Score because both classes have the same weight

## Preparing environment

import pandas as pd
import numpy as np
import sys
from sklearn.metrics import f1_score, make_scorer
from sklearn.model_selection import train_test_split, RandomizedSearchCV, StratifiedKFold, cross_val_score
from xgboost import XGBClassifier
import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import TensorDataset, DataLoader, Subset
import optuna
sys.path.append('../high_performance_employee_resign_prediction')
from utils import paths

import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

## Importing data

train_df = pd.read_csv(paths.data_processed_dir('reduced_train.csv'))
test_df = pd.read_csv(paths.data_processed_dir('reduced_test.csv'))

# Saving id_employee_employee for submission

id_col = test_df['id_employee']

y = train_df['resign']

# Dropping unnecessary columns

X = train_df.drop(columns=['id_employee', 'id_last_boss', 'resign'])
X_test = test_df.drop(columns=['id_employee', 'id_last_boss'])

categorical = ['gender', 'gender_boss', 'seniority', 'recruitment_channel', 'join_year', 'marital_estatus', 'join_month_boss', 'join_year_boss', 'performance']
X[categorical] = X[categorical].astype('category')
X_test[categorical] = X_test[categorical].astype('category')

X.columns

## Preparing cross validation

cv = StratifiedKFold(n_splits=5, random_state=42, shuffle=True)

## Evaluating XGBoost Model

# Creating base model

xgb_base = XGBClassifier(random_state=42, eval_metric='logloss', objective='binary:logistic', booster='dart', tree_method='hist', enable_categorical=True, device='cuda')

xgb_base.fit(X, y)

# Making cross-validation in xgboost base

cv_results = cross_val_score(xgb_base, X, y, cv=cv, scoring='f1')
cv_results

Initial f1 results are good, let's optimize the model with cross-validation

# Defining param grid

param_dist = {
    'n_estimators': np.arange(100, 1100, 100),
    'max_depth': np.arange(3, 11),
    'learning_rate': np.logspace(-3, 0, 10),
    'subsample': np.linspace(0.5, 1.0, 6),
    'colsample_bytree': np.linspace(0.5, 1.0, 6),
    'reg_alpha': np.logspace(-3, 1, 10),
    'reg_lambda': np.logspace(-3, 1, 10),
    'min_child_weight': np.arange(1, 6)
}

# Defining random search

random_search = RandomizedSearchCV(
    estimator=xgb_base,
    param_distributions=param_dist,
    scoring=make_scorer(f1_score),
    n_iter=50,  # Number of different combinations to try
    cv=cv,
    verbose=2,
    random_state=42,
    n_jobs=-1  # Use all available cores
)

# Optimizing model

random_search.fit(X, y)

print(f"Best parameters found: {random_search.best_params_}")
print(f"Best cross-validation F1 score: {random_search.best_score_}")

xgb_params = random_search.best_params_

best_xgboost = XGBClassifier(random_state=42, eval_metric='logloss', objective='binary:logistic', booster='dart', tree_method='hist', enable_categorical=True,
                             **xgb_params)

# Fitting the model 

best_xgboost.fit(X, y)

# Predicting values

y_pred_xgb = best_xgboost.predict(X_test)

# Saving predicted values

sub_xgb = pd.DataFrame(y_pred_xgb, columns=['abandono_6meses'])
sub_xgb = pd.concat([id_col, sub_xgb], axis=1)
sub_xgb.rename(columns={'id_employee': 'ID'}, inplace=True)
sub_xgb.to_csv('../results/sub_xgb_rfe_manual.csv', index=False, sep=',')

Kaggle Score: 0.6106
